Loading animation works
It's loaded using the 'onBeforeAction()' in IronRouter. This means it runs before any page loads. It's only a basic implementation of it so far, and should be adjusted per view that gets loaded--most likely with the 'onBeforeAction()' for those specific routes.

For reference: https://www.eventedmind.com/classes/how-to/meteor-how-to-use-an-ios7-style-loading-indicator-with-iron-router