Python script execution on the server is possible now
>>Added: meteorhacks:npm
You will have to provide your own python script, and change the "callPy" to point to your script. But when those are hooked up, you should see the output from the Python script on the server side terminal. Keep in mind, it would NOT take much more work (change the string) to execute anything server side, not just Python.